+++ lib/still_active/workflow.rb
@@ -5,6 +5,7 @@ require_relative "gitlab_client"
 require_relative "repository"
 require_relative "../helpers/health_score_helper"
 require_relative "../helpers/libyear_helper"
+require_relative "../helpers/ruby_helper"
 require_relative "../helpers/version_helper"
 require "async"
 require "async/barrier"
@@ -15,14 +16,31 @@ module StillActive
   module Workflow
     extend self
 
-    def call
+    def call(&on_progress)
       task = Async do
         barrier = Async::Barrier.new
         semaphore = Async::Semaphore.new(StillActive.config.parallelism, parent: barrier)
         result_object = {}
+        total = StillActive.config.gems.size
+        completed = 0
         StillActive.config.gems.each_with_object(result_object) do |gem, hash|
           semaphore.async do
-            gem_info(gem_name: gem[:name], result_object: hash, gem_version: gem[:version])
+            gem_info(
+              gem_name: gem[:name],
+              result_object: hash,
+              gem_version: gem[:version],
+              source_type: gem[:source_type] || :rubygems,
+              source_uri: gem[:source_uri],
+            )
+          rescue Octokit::TooManyRequests
+            $stderr.print("\r\e[K") if on_progress
+            $stderr.puts("rate limited checking #{gem[:name]}: set GITHUB_TOKEN to increase your limit")
+          rescue StandardError => e
+            $stderr.print("\r\e[K") if on_progress
+            $stderr.puts("error occurred for #{gem[:name]}: #{e.class}\n\t#{e.message}")
+          ensure
+            completed += 1
+            on_progress&.call(completed, total)
           end
         end
         barrier.wait
@@ -31,13 +49,35 @@ module StillActive
       task.wait
     end
 
+    def ruby_freshness
+      RubyHelper.ruby_freshness
+    end
+
     private
 
-    def gem_info(gem_name:, result_object:, gem_version: nil)
-      result_object[gem_name] = {}
+    def gem_info(gem_name:, result_object:, gem_version: nil, source_type: :rubygems, source_uri: nil)
+      result_object[gem_name] = { source_type: source_type }
       result_object[gem_name][:version_used] = gem_version if gem_version
 
-      vs = versions(gem_name: gem_name)
+      case source_type
+      when :path
+        gem_info_path(gem_name: gem_name, result_object: result_object)
+      when :git
+        gem_info_git(gem_name: gem_name, result_object: result_object)
+      else
+        gem_info_rubygems(
+          gem_name: gem_name,
+          gem_version: gem_version,
+          result_object: result_object,
+          source_uri: source_uri,
+        )
+      end
+
+      result_object[gem_name][:health_score] = HealthScoreHelper.gem_score(result_object[gem_name])
+    end
+
+    def gem_info_rubygems(gem_name:, gem_version:, result_object:, source_uri:)
+      vs = versions(gem_name: gem_name, source_uri: source_uri)
       repo_info = repository_info(gem_name: gem_name, versions: vs)
       commit_date = last_commit_date(
         source: repo_info[:source],
@@ -89,9 +129,33 @@ module StillActive
           ),
         })
       end
-      result_object[gem_name][:health_score] = HealthScoreHelper.gem_score(result_object[gem_name])
-    rescue StandardError => e
-      $stderr.puts "error occurred for #{gem_name}: #{e.class}\n\t#{e.message}"
+    end
+
+    def gem_info_git(gem_name:, result_object:)
+      repo_info = repository_info_from_installed_gem(gem_name: gem_name)
+      commit_date = last_commit_date(
+        source: repo_info[:source],
+        repository_owner: repo_info[:owner],
+        repository_name: repo_info[:name],
+      )
+      archived = repo_archived?(
+        source: repo_info[:source],
+        repository_owner: repo_info[:owner],
+        repository_name: repo_info[:name],
+      )
+      scorecard = DepsDevClient.project_scorecard(project_id: repo_info[:project_id])
+
+      result_object[gem_name].merge!({
+        repository_url: repo_info[:url],
+        last_commit_date: commit_date,
+        archived: archived,
+        scorecard_score: scorecard&.dig(:score),
+      })
+    end
+
+    def gem_info_path(gem_name:, result_object:)
+      repo_info = repository_info_from_installed_gem(gem_name: gem_name)
+      result_object[gem_name][:repository_url] = repo_info[:url]
     end
 
     def fetch_deps_dev_info(gem_name:, version:)
@@ -106,11 +170,38 @@ module StillActive
       }
     end
 
-    # makes network request
-    def versions(gem_name:)
-      Gems.versions(gem_name)
+    def versions(gem_name:, source_uri: nil)
+      if github_packages_uri?(source_uri)
+        fetch_github_packages_versions(gem_name: gem_name, source_uri: source_uri)
+      else
+        Gems.versions(gem_name)
+      end
     rescue Gems::NotFound
       []
+    end
+
+    def github_packages_uri?(uri)
+      uri.is_a?(String) && uri.include?("rubygems.pkg.github.com")
+    end
+
+    def fetch_github_packages_versions(gem_name:, source_uri:)
+      base = URI(source_uri.chomp("/"))
+      namespace_path = base.path
+      path = "#{namespace_path}/api/v1/gems/#{gem_name}/versions.json"
+      token = StillActive.config.github_oauth_token
+      headers = token ? { "Authorization" => "Bearer #{token}" } : {}
+      HttpHelper.get_json(base, path, headers: headers) || []
+    end
+
+    def repository_info_from_installed_gem(gem_name:)
+      valid_repository_url =
+        installed_gem_urls(gem_name: gem_name).find { |url| Repository.valid?(url: url) }
+      repo = Repository.url_with_owner_and_name(url: valid_repository_url)
+      project_id = if repo[:url]
+        host = repo[:source] == :gitlab ? "gitlab.com" : "github.com"
+        "#{host}/#{repo[:owner]}/#{repo[:name]}"
+      end
+      repo.merge(project_id: project_id)
     end
 
     def repository_info(gem_name:, versions:)